move deps into nested folder
